R Programming assignment 2_ CachingMatrix function
Two functions. makeCacheMatrix(matx) creates an object with two variables and outputs a list of functions, which can later be accessed over the list operator.
cacheSolve(obj) takes the obj of type makeCacheMatrix and either inverts the matrix saved in the object, or outputs the already saved inverse.
Changes to the older version add in the comments, where and why the "<<-" operator was used.